crypto: fix generation of F1Field class declaration in typing/ffjavascript
For some weird reason, when the gen-types.sh script calls tsc to
automatically generate type declarations for ffjavascript, the result
is not always including the `extends` inheritance from the
implementation of the F1Field class (a.k.a. ZqField).

This means that the constructor declaration is missing, which in turn
causes tsc compilation failures when instantiating that class.

So until we can figure out how to persuade tsc to include `extends` in
the right place, add an ugly hack to gen-types.sh which monkey-patches
the code to include it.