rename advisor.adviceReward() to .getRewardAdvice()
"Advice" is a noun, therefore "adviceReward" is a combination of two
nouns which doesn't entirely make sense.  "Advise" is a verb, but
"adviseReward" would mean the reward is the thing getting advised,
whereas it's actually the caller which is getting advised, not the
reward.  So change to "getRewardAdvice" which is clearer.