add git commit message template
Without this patch, we are regularly merging commits with
low quality commit messages.

This is a problem because:

- It makes code review much harder.

- It makes it harder to perform history manipulation operations such
  as rebases, reorderings, fixups, cherrypicks, merges, conflict
  resolution etc.

- It makes it harder to understand why/how changes were made
  at any point in the future, e.g. when using git blame.

- It fails to encourage the discipline of ensuring that each
  commit is an atomic unit of logical change, with a high degree
  of focus.

- It makes it easy to mix multiple changes within a single commit,
  e.g. refactoring with bug fixes or new features.

- It makes it easy to split a single logical change across multiple
  commits.

This patch solves the problem by adding a commit message template
(which this commit itself uses) providing a clear structure for the
message, together with a checklist of items to remember, and useful
resources for understanding best practices more deeply.  When this
structure and advice is followed, it should guarantee much better
commit quality.

Issue-URL: n/a

The message template is "installed" (configured) automatically
by running "yarn" or "yarn install".