VotingPowerInformer contract added
The amount of $ZKP tokens a voter has on stake with the Staking contract defines the voting power of the voter.
A voter (staker) may delegate his `own` power to another voter.
To let "consumers" count votes with a method they may require, the Staking contract provides `own` and `delegated` voting power separately.
It creates a problem for plugins which the pantherprotocol namespace on the snapshot.org uses:
these plugins can not directly apply a pair of values instead of the single resulting value.

The VotingPowerInformer is a contract with "view" only functions.
Being called by a plugin, it reads voting power data from the Staking contract and converts it to the format the plugin may use. The contract supports "ordinary" and "quadratic" voting.