add default MR template
Without this patch, merge requests are being submitted with
various quality issues, e.g. combining totally unrelated changes
in a single MR, or not providing an explanation of why the change
is needed and why it's the best approach.

This is a problem because it makes code review much harder and slower,
leading to code review backlogs.  So MRs stay open for longer, which
in turn leads to merge conflicts and more wasted time.  It also
creates technical risk when changes are not clearly articulated, since
increased complexity means that unintended or unnoticed side effects
can easily slip past the reviewer.

So provide a default MR template which encourages a clear description
of the change, and provides a quality checklist to ensure the key
points are achieved.

See https://docs.gitlab.com/ee/user/project/description_templates.html
for more info on MR templates.