MaticRewardPool.sol drafted (WIP)
Tokens from the Reward Pool are supposed to be bridged to Matic at once, rather than "minted" when needed, as it is on the mainnet. Furthermore, no VestingPools instance will run on Matic.
Therefore, the "mainnet" version of the RewardPool contract, that interacts with the VestingPools directly, can not be used.
This commit introduces the version of the RewardPool contract designed to work on Matic.