contracts: enhance StakesReporter unit tests
Change the behaviour of getStakesInfo() so that the unclaimedRewards
array is always the same length as the stakes array; even when it
contains claimed stakes, the rewards values should be present and
zero.

Also try to make tests a bit more realistic, and further from just
reimplementing the exact same logic in the contract.