avoid broken ClickUp / GitLab integration
Without this patch, the .gitmessage template includes a "magic"
reference to a ClickUp task, which ClickUp will automatically
detect and change the task to COMPLETED status.

This is a problem because it happens whenever the commit is pushed
to GitLab, rather than only when the commit is merged to main as
part of a merge request.

This patch solves the problem by removing the magic line from
the .gitmessage template.

Adam has had a meeting with ClickUp and reported that this integration
needs to be improved.  The representative agreed to pass on the
feedback to their product team.