contracts: hardhat task to fetch unclaimed rewards
Without this patch, we could not get the list of all stakes and accumulated rewards.

This patch solves the problem by creating the hardhat task to fetch unclaimed rewards.
Depending on the chainId, the task uses either RewardMaster contract on the Ethereum
mainnet or StakesReporter on Polygon. this task uses 2 files generated by the
staking:list task: 1) stakes claimed JSON file and 2) stakes created JSON file. The
difference between these 2 files is the list of stakes that are not claimed yet. Then,
for an unclaimed list of stakes, the current task fetches the rewards for each stake
from the corresponding contract.