contracts: timelock in StakeRewardController2.sol::rescueErc20
On withdrawal by the owner of accidentally sent (i.e. stuck) tokens from the StakeRewardController2 contract with the `rescueErc20` methos, in case the token to withdraw is the reward token, the method reverts before 2022-08-15T00:00:00.000Z.
Before this commit, the method ensured if the balance if enough to pay `unclaimedRewards`, which was less flexible and resulted in more risks of freezing the reward token with the contract forever.