dapp: fix increased size of ciphered.data from 48 to 80 bytes
Without this patch, ciphered.data has a size of 80 bytes (instead of 48) which the smart contract cannot
read. The smart contract splits passed data into chunks of the fixed size. Passing one of the chunks with
incorrect size breaks the execution and, therefore, the smart cotnract reverts the transaction.

This patch solves the problem by decreasing the size of ciphered.data from 80 to 48 bytes by using
Uint8Array as inputs instad of hex. Passing the input string as a hex into cypher increases the size of
the encrypted data for obvious reasons.

Issue-URL: https://app.clickup.com/t/2dfzxx8