contracts: download NewCommitments events from PantherPool contract
Just for advanced staking, we need to periodically build the MASP v0
Merkle tree and make it available for download by the dApp, for users to
be able to activate the early redemption feature. The tree consists of
the leaves of commitments from NewCommitments events emitted by the
PantherPool contract. To prepare for building of the tress, first, we
need to download the NewCommitments from the PantherPool contract.

This patch solves the problem by implementing a hardhat task that
fetches NewCommitments events from the PantherPool contract. This task
requires several input arguments:

- address - Pool contract address;
- start - Starting block number to look from;
- out - direcotry to write to;
- chunks-prefix - prefix of files to write chunks to. Downloaded chunks
  will be read on the next run reducing the amount of the requests to
  the chain
- end (optional) - ending block number to look to.

Issue-URL: https://app.clickup.com/t/2au3w9z