crypto: removing CONSTANTS from triad-merkle-tree package
Without this patch, there is CONSTANTS file containing constant variables used in and related to preZKP. These
variables are not needed across the whole workspace and are mostly needed in tests for the specific pre-built trees.

This patch solves the problem by removing the CONSTANTS file from the triad-merkle-tree package and adding constants
directly in places where they are needed. Namely, in test and utils files.