Modify the `adviser:add` to add adviser for either classic or advanced staking.
This is a problem because currently we have 2 type of adviser and based on some requirements we add one of them. the current task structure is not flexible enough to address this issue.

This patch solves the problem by adding 3 parameters for the task: `replace`, `classic`, and `mainnet`.
if the `replace` is `true`, it means the `rewardMaster` already has an adviser and we aim to replace it.
if the `classic` is `true`, it means we aim to add an adviser for classic staking otherwise, the advanced adviser will be added.
if the `mainnet` is true, it means we aim to deploy an adviser for mainnet or its test networks (Kovan, Ropsten, etc), otherwise, the adviser for polygon or its test network (Mumbai) will be added.

Hardhat expect to see no error before start compiling. when we import `types/contracts` in `lib/staking`, it means we imported something that does not exist and since `lib/staking` is imported in one of the hardhat tasks, hardhat complains. to fix it, `TS_NODE_TRANSPILE_ONLY` env is set to true before executing any hardhat task.

Issue-URL: https://app.clickup.com/t/2chuc8x