contracts: fix of scaling in PantherPoolV0::_processDepositedAsset
Before this update, the code of the _processDepositedAsset function did not check if the
extAmount is scalable w/o a reminder (i.e. the extAmount is a multiple of the scale).
It is a problem since the remainder, if it takes place, gets locked in the Vault forever.
This commit adds the code, which reverts if the scaling of the "external" amount into
the "internal" amount (i.e. the amount inside zk-circuits) is enabled for an asset, and
the external amount can not be scaled w/o a reminder.