Add a task to set Zkp rewards limit Without this patch, there is no task to set Zkp rewards limit for `AdvancedStakeRewardController` contract.
This is a problem because the rewards in `AdvancedStakeRewardController` won't be allocated to stakes automatically. first zkp tokens should be transfer to the contract. then call `setZkpRewardsLimit()` to allocate rewards. this function is also approve Vault to spend staked token on behalf of contract.

This patch solves the problem by adding a task to `tasks/rewards-limit-add.ts` to invoke `setZkpRewardsLimit()`

Issue-URL: https://app.clickup.com/t/2f1vk4r