Rename `SNARK_FIELD_SIZE` to `BN254_FIELD_SIZE` in the whole project
Without this patch, we won't be in sync with latest changes & developments in the backend side of this project.

This is a problem because the backed team decided to adopt another name for some of their coding sections in the backend, so the new naming should be applied semantically speaking.

This patch solves the problem by renaming `SNARK_FIELD_SIZE` to `BN254_FIELD_SIZE` in the whole project.

Issue-URL: https://app.clickup.com/t/2ff4why