DRY summing of BigNumbers via a new helper function
Without this patch, we lack a handy helper function to accumulate arrays of ether.BigNumber.

This is a problem because we have duplicated this code in multiple files in this project.

This patch solves the problem by DRYing those multiple use cases by creating a new helper function file in dapp/src/lib folder and reusing it in above occurances.

Issue-URL: https://app.clickup.com/t/2gpxzdg