amplify.yml: add missing COMMITMENT_TREE_URL_80001
Without this patch, AWS Amplify deployments are always missing
the COMMITMENT_TREE_URL_80001 environment variable, because that
variable was added in 13d61ea9 but not to amplify.yml.

This is a problem because even if it's configured correctly in
Amplify, it will be impossible for the dapp to download advanced
staking Merkle trees.

This patch solves the problem by adding it to amplify.yml.

Issue-URL: https://app.clickup.com/t/2hkqhf3

ClickUp: #2hkqhf3