contracts: PrpGrantor::issueOwnerGrant added, MAX_PRP_GRANT decreased
PrpGrantor implements the `issueGrant` method.
This method logic, which is acceptable and assumed to be in the V1 also, has the following issues.
- The amount of a grant is fixed and defined by the grant "type"
- The grant "type" needs to be registered before calling the method
- The total amount of grants of this type, however, is not limited (other than by MAX_PRP_GRANT)
This logic does not work well enough in some cases, for example:
- Total amount of grants need to be limited by some specific value
- The grant is supposed to be issued by the owner, preferably, with a single transaction
(And these are exactly the cases for the "advance staking" rewarding in PRP)

This commit introduces the method `issueOwnerGrant` with works better in the above said cases.
The commit updates unit tests (and a mock contract used in tests).

Another update this commit introduce is reducing of the MAX_PRP_GRANT which is too large.