Define a PrpGrantor contract address and a fixed amount of PRP per stake in the advanced controller. Without this patch, the `AdvancedStakeRewardController` calls the external `PantherPool` and NFT contract to grant PRP and mint NFT to itself to let `Vault` transfer them from `AdvancedStakeRewardController` on every stake. Staker can unlock them later and withdraw from `Vault`.
This is a problem because the `PrpGrant` has become a stand-alone contract and its functionality cannot be accessed from `PantherPool`. Also, generating PRP (and also NFT) for each stake can affect the PRP value because of not have a defined cap value.

This patch solves the problem by defining a limit for PRP and NFT rewards. The contract does not issue/mint PRP/NFT to itself anymore. The rewards will be issued/minted to the contract and the contract will be able to allocate them to the user.
The `PRP_GRANTOR`, `PRP_REWARD_PER_STAKE`,  `prpRewardsLimit` and,  `nftRewardsLimit`  have been added to the contract. The Multisig wallet issues PRP and mint NFTs to the contract. Then PRP and NFT (as well as ZKP) reward limits can be filled by anyone.
The following functions have been added to set PRP and NFT reward limits:
- `setPrpRewardsLimit`
- `setNftRewardLimit`

The logic for approving `Vault` has been placed in a separate function:
`increaseVaultAllowance`

A function has been added to set limits for all kinds of rewards and also approve the `Vault`:
`prepareRewardsLimit`

Issue-URL: https://app.clickup.com/t/2jej3z9