crypto: add declaration map for navigation to definition
Without this patch, when you navigate to a definition, the source code
is not shown. Instead, the compiled type declaration is displayed.

This is a problem because it results in a bad developer experience as
code navigation does not work.

This patch solves the problem by adding a declaration map for code
navigation by changing the config settings of the @crypto workspace.

Issue-URL: https://app.clickup.com/t/31fy9dj