dapp: add ZAssetsRegistry contract to fetch zAssetsId in MASP v0.5
Without this patch, zAssetsId was fetched from the pool contract. With
the version 0.5 of the MASP smart contracts, the zAssetsId is fetched
from the ZAssetsRegistry contract.

This patch solves the problem by adding the ZAssetsRegistry contract to
the dApp: ABI, environmental variables, and getter methods. As the
functionality of the Pool contract changed, the patch updates the ABI of
the pool contract. Finally, exit() function is updated to use
ZAssetsRegistry contract instead of Pool contract.

Issue-URL: https://app.clickup.com/t/3461ybm