contract:  make rewarding times/APYs mutable
Without this patch, the rewarding times and APYs are immutable.

This is a problem because we won't be able to reuse the contract to setup another advanced staking cycle.

This patch solves the problem by adding a structure that stores those 4 rewarding parameters in one EVM storage slot.
The private `_setRewardParams()` has been added to set the rewarding parameters. This function can be called via the constructor or external `updateRewardParams()` that can only be called by the owner.

Issue-URL: https://app.clickup.com/t/2k0c4jv