dapp: check if the user is blocked
Without this patch, we can't know if the user is allowed or not allowed to use the app due to the country of residence.

This is a problem because we need to know how is and who isn't allowed to use the application.

This patch solves the problem by adding functionality to get the country of residence for the user, and check if this country is blocked or not.

Issue-URL: https://app.clickup.com/t/34bpbcy