Revise "build:types" in package.json
Without this patch, the gitlab pipeline fails for this task because
sometimes it needs to build:types before linting the codebase.

This is a problem because for this task, we first need to run
yarn build:types in order to get access to the build files in crypto/lib
folder.

This patch solves the problem by revising "build:types" in package.json
file accordingly

Issue-URL: https://app.clickup.com/t/35tbwcv