dapp: PrpGrantor to request unused PRP grant on any network
Without this patch, the PRP grant amount is not available to the
networks except Mumbai or Polygon.

This is a problem because the user would like to stake through the
bridge but he/she doesn't know if the PRP rewards are available.

This patch solves the problem by calling PrpContract to get the PRP
grant amount from any network. More specifically, we initialize the
PrpGrantorContract with the providers of the MASP chains. MASP chains
are Mumbai, Polygon, or Hardhat only specified via added new type. Thus,
the contract is initialized using the provider of one of the MASP chains
and as a result, can request the PRP grant amount from any network.
Also, this patch adds the required environmental variable MASP_CHAIN_ID
that specifies the network the PrpGrantorContract calls.

Issue-URL: https://app.clickup.com/t/35td9x2