contracts: Modify `adviser:add` task Without this patch, there will be some situations that have not been covered by the task:
-- Adding adviser for advanced staking on Mainnet.
-- Adding msgRelayer as the second oracle to polygon's rewardMaster

This patch solves the problem by retrieving the
 mentioned contracts based on the task flag:

flag `--relayer` defines if the msgRelayer is going to be added as the oracle. otherwise,
 the staking contract.
flag `--mainnet` defines if the adviser will be added to rewardMaster on mainnet,
otherwise, polygon.
flag `--classic` defines if the adviser will be added to for classic staking type,
otherwise, advanced staking.
flag `--replace` defines if the adviser is going to replace the current adviser.

The task's parameters have been replaced with flags to have a cleaner code.
 As the type of all of them was boolean, it does make more sense to use flag simply.
 In this case, we are not forced to pass the flags as they have a default `false` value.

Issue-URL: https://app.clickup.com/t/2qewz0r