contracts: fix 'getContract' does not exist on type ethers
Without this patch, there is a bug in hardhat-deploy-ethers of version
0.3.0-beta.12 that causes hardhat to throw an error that property
'getContract' does not exist on type ethers & HardhatEthersHelpers'.

This patch solves the problem by upgrading the version of
hardhat-deploy-ethers to 0.3.0-beta.13. Also, the patch updates the
import paths of @nomiclabs/hardhat-ethers to match a new version and
ignores the bug that does not allow accessing the filter as function.

Issue-URL: https://app.clickup.com/t/3b51aw4