dapp: feat fetching exit delay time from the Pool contract
Without this patch, the exit delay time is set in the .env config file.

This is a problem because it requires unnecessary manual intervention to
update the config file when the exit delay time is changed in the Pool
contract. Moreover, the exit delay time also needs to be updated in the
environmental variables of AWS amplify.

This patch solves the problem by fetching the exit delay time directly
from the Pool contract. To accomplish that ABI of the Pool contract was
updated to the latest that has this function. Then, the service function
getExitDelay() was created that requests the Pool Contract to get the
exit delay time. Finally, this patch updated the slice with the
parameters of the Pool Contract to accommodate a new variable. The patch
removed the old use of the environmental variable.

Issue-URL: https://app.clickup.com/t/3dn9vk6