dapp: fix hot reload
Without this patch, we will not have hot reload working. I had to
upgrade webpack and add a few options to make this working again.

This is a problem because recent updates of storybooks and webpack cause
that HMR stoped working

This patch solves the problem by updating webpack to proper version,
adding a few more options to allow app to hot reload changes

Issue-URL: https://app.clickup.com/t/3gpdvb5