root: refactoring of lib/keychain logic
Without this patch, we have duplication of keychain logic.

This is a problem because we must maintain two different code versions
of the same logic, which are prone to diverge.

This patch solves the problem by refactoring the code to use a joint
code base and extracting the code into the shared workspace
@panther-core/crypto.