root: refactor moving services/keychain to crypto/sdk/keychain
Without this patch, the keychain service is not able to be used by the
contracts workspace. This patch moves the keychain service to the
crypto/sdk/keychain. Then, the SDK keychain functions are used to derive
the keys in tests. Also, the tx error handling is moved to the crypto
workspace.

Issue-URL: https://app.clickup.com/t/3kqrj3c