crypto: refactor move parseTxErrorMessage() in dApp services
Without this patch, the parseTxErrorMessage() is defined in the @crypto
workspace. This function is the @dApp workspace specific, so it should
be defined there. This patch moves the function to the @dApp workspace.
Also, this patch creates generateRootKeypairs() function in the dApp
workspace to make the abstract boundaries more clear.