Slither vulnerability - incorrect-equality
Without this patch, Slither will complain about incorrect-equality
vulnerability.

This is a problem because this vulnerability is positive only when we
use the strict equality to determine if an account has enough Ether or
tokens. So these are false positive.

This patch solves the problem by silencing the warning for particular
lines which are the cause of this vulnerability.

Issue-URL: https://app.clickup.com/t/2wv124q