crypto: refactor Merkle tree
This patch adds the function quadLeafIndexRangeForTreeId, which returns
the range of the quad leaf indices for the specific tree ID. This is the
number of quad leaves the user can add to a tree before it is split into
a new tree. Also, the patch adds new frequently used constants of tree
size and leaf zero value. Moreover, the patch renames some functions to
have more explicit names and adds new tests.