dapp: feat build MerkleTree in the browser
Without this patch, we build the tree using CLI scripts and upload the
resulting JSON to the AWS S3 bucket.

This is a problem because it requires much manual work, and every
additional advanced stake requires a new version of the tree.

This patch solves the problem by retrieving the commitments from the
randomly chosen subgraph URL and building the Merkle tree in the
browser. Also, this patch removes code associated with the downloading
tree from the s3 bucket. Moreover, this patch cleans the required
variables in env.ts.

Issue-URL: https://app.clickup.com/t/3ufra4d