contracts: fix various issues in the bridge deployment process
Without this patch, there are some issues with deploying the
bridge-related contracts:
- The msgSender script skips process where the network is mainnet/goerli,
that is wrong. It should be skipped if the network is polygon/mumbai.
- The `getContractEnvAddress` function gets the env variable names
based on the network, however in the bridge deployment process, we need
the variable of the opposite network. i.e, if we are deploying the
`msgSender` on the mainnet, we need the address of relayerProxy on the
opposite network, which is polygon. The same issue is found in
`relayerImp` process. the contracts needs to be deployed on polygon and
it needs the address of msgSender on the opposite network.
- There are also some issues in the logged messages.