contracts: fix pnft deployment script for prod/dev
Without this patch,
- The script for setting the minter for pnft is
executed for both prod and dev environments, but it must be executed
only in dev environment.
- In dev environment, If the multisig address is defined as env file,
then it is used as owner of pnft. In this case, it is impsossible to
use this address to send tx for changing the pNft owner.

This patch solves the problem by:
- Adding the `isProd()` helper function and use it when setting the
owner of pNft, so the script is skipped in production.
- In dev environment, Using the deployer as pNft owner, then it's possible to
change the pNft owner on behalf of deployer.

Issue-URL: https://app.clickup.com/t/320ugjv