dapp: fix replace subgraph name with hash id
Without this patch, in our requests to the subgraph, we used account
names in the link. This is a problem because the malicious subgraph
owner can change the subgraph under the same name and break our dApp.
This patch solves the problem by replacing the request to use the
subgraph's hash id instead of the subgraph owner name.