Changing circuit folder from javascript to typescript.
Without this patch, circuit will use javascript and it will miss out
typescript benefits.

This is a problem because circuit folder will be inconsistent with other
project folder strcuture which uses typescript.

This patch solves the problem by converting the circuit folder to
typescript. This patch will solve the timeout issue in the existing test
cases and includes convertion of the existing test cases to typescript.

Issue-URL: https://app.clickup.com/t/85zrgf0ma