feat(dapp): optimize performance of UTXO update
Without this patch, We will face a performance bottleneck with a big UTXO set with
the risk of being blocked from using the RPC node and slow UI update.

This is a problem because we are iterating through the list of UTXOs and
making a call to check if the nullifier is spent for each UTXO. This is far from
optimal and makes too many requests to the RPC node with a risk of being blocked
due to spam.

This patch solves the problem by bundling all RPC calls into one single HTTP request
transaction using ethers.multicall library

Issue-URL: https://app.clickup.com/t/866949buy