feat(dapp): general code splitting
Without this patch, our dapp entry point will be big (6.51MB)

This is a problem because it will take longer for the browser to load the dapp.

This patch solves the problem by applying code splitting on our bundle and make
it as the following. `main.js` Our code base (node_modules + our code)
`circomlib.js` will be cached. `vendor-react.js` will be cached.

Issue-URL: https://app.clickup.com/t/86699e1yj