feat(dapp): add env variables for staging and production
Without this patch, we cannot conditionally build and render the app for
staging and production.

This is a problem because we need to add the testing background image
for staging, and, most likely, in the future, we will need to have some
other distinct features for staging and production.

This patch solves the problem by adding the
TARGET_ENV={development,staging,production} env variable. As a result,
this patch updates the webpack config to build the app for staging and
production conditionally: more specifically, it conditionally adds the
source map files, optimization of the js files, and Sentry build upload.
Moreover, the patch updates package.json, the Amplify and GitLab scripts
to use the new env variable. Also, the patch adds an env.staging file
for test staging deployment and updates the env example files. Finally,
the patch updates the GitLab CI script, removing the unnecessary second
run of the dapp test.