ci(root): add max changed files rule to dangerjs automation
Without this patch, there is no limit on changed files in one merge
request.

This is a problem because it makes it easy to commit a large number of
files which, in turn, is hard to review.

This patch solves the problem by adding a rule to dangerjs which will
warn the assignee if the number of changed or newly added files exceeds
20.