fix(root): dangerjs test rules for modified files
Without this patch, the dangerjs test rules for modified files do not
work. Instead, the script only checked newly created files.

This is a problem because some tests could be updated or added to the
existing files.

This patch solves the problem by checking if any test files are present
in the modified files.