fix(dapp): review eslint warning
Without this patch, an eslint warning will be shown when running `yarn lint:ts`

This is a problem because the warning says that the current version of eslint
doesn't support typescript 4.9.4 that we just updated it recently.

This patch solves the problem by upgrading all `eslint` related packages into
the latest version and resolving all the errors caused by the new version.

Issue-URL: https://app.clickup.com/t/86779bvz1