docs(dapp): add E2E tests docs
Without this patch, we will not have a documented way to run the E2E tests for
the dApp.

This is a problem because there is required environment variables to run the
tests and tests structure stander we need to follow.

This patch solves the problem by adding `dapp-e2e.md` file to document how to
run the tests, required environment variables and the `dapp/cypress/` folder
structure

Issue-URL: https://app.clickup.com/t/8669jq330