ci(root): optimize CI/CD pipeline
Without this patch, the cache in the pipeline is not properly set. As a
result, all jobs we ran without using an existing cache took
substantially longer to complete.

This patch solves the problem by setting the cache policy correctly.
Also, the patch added logging of the timings of the tasks.