fix(contracts): update typescript
Without this patch, we can't use the latest version of typescript in the
`contract` workspace because we were getting the error below

"Error: Debug Failure. False expression: Non-string value passed to
`ts.resolveTypeReferenceDirective`, likely by a wrapping package working with an
outdated `resolveTypeReferenceDirectives` signature. This is probably not a
problem in TS itself. "

This is a problem because the goal is to have the same typescript version across
all workspaces.

This patch solves the problem by...
1. Use the typescript version like the one in `yarn.lock` and other workspaces
2. Updating `hardhat` and other `hardhat-related` packages to fix the error
above.
3. Update the `yarn test` script to remove the `artifacts/` and `cache/`
directories before starting the tests as it was interfering with the tests and
causing them to fail.

Issue-URL: https://app.clickup.com/t/865bm8agq