fix(circuits): configuration fixes for circuits folder
Without this patch there is a minor bug which does not test the
environment variable availability and test cases control flow will be
misleading.

This is a problem because test cases will run in docker environment even
if we choose to run tests in local environment.

This patch solves the problem by importing the environment variables and
fixing the configurations in package.json. Same has been updated in the
documentation as well.

Issue-URL: https://app.clickup.com/t/85zrgkazu