ci(root): pipeline time-saving optimization
Without this patch, the CI pipeline has 6 tasks executed in parallel.
This consumes substantial pipeline time, especially when the pipeline is
performed without a cache. This patch reduces the number of tasks to 2
making overall job execution longer (from 10 to 15 minutes). However,
this saves more than 70% of the pipeline time and saves costs.