refactor(dapp): add Web3Provider type
Without this patch, we will use the type any e are referencing the `library`
object (aka provider).

This is a problem because `any` type is inherently unsafe and it is a bad
practice.

This patch solves the problem by adding the right type for the `library` object
which is `Web3Provider`. Note: We had to update prettier because it was throwing
this error (`SyntaxError: '=' expected.`) whenever trying to make a commit.
After some research truns out the current version of prettier is not compatible
with our TS version.  Check out "why I use "import type" tell me SyntaxError?"
https://github.com/prettier/prettier-vscode/issues/2563

Issue-URL: https://app.clickup.com/t/8669wjukg