ci(dapp): drop amplify ci deployment
Without this patch, we will still use Amplify to deploy the dApp from the CI/CD.

This is a problem because we no longer need it as the `ifps-deploy.ts` script is
doing the same thing with much more advanced features and a real environment
that is similar to the production one (IPFS).

This patch solves the problem by updating `.gitlab-ci.yml` file and removing the
`preivew_script` that is used to deploy the dApp into amplify. This patch does a
cleanup to the `syncEnv` script to remove Amplify from it as well.