feat(dapp): upgrade react router to v6
Without this patch, we are using old
version of react router in our dapp.

This is a problem because we must be
up-to-date with latest version.

This patch solves the problem by
installing the latest version of
react-router & react-router-dom
and then updating the App.tsx
and routes.ts files accordingly.

Issue-URL: https://app.clickup.com/t/861mmgzqp