feat(dapp): implement IndexedDB for persisted storage
Prior to this update, the dapp persisted the wallet state in local
storage, which is inaccessible from Web workers, making IndexedDB a more
suitable option. The dapp will utilize JSON for data formatting and
IndexedDB for file storage. Data is classified into four groups based on
its sensitivity and recreatability:

1. Sensitive and recreatable data: Not stored to prevent security
   breaches.
2. Sensitive but not recreatable data: Encrypted using public key
   encryption before storage.
3. Non-sensitive and non-recreatable data: Stored without encryption.
4. Recreatable and non-sensitive data: Cached or not stored for
   performance optimization.

This update implements the described storage mechanism for the
wallet/UTXOs, ui/acknowledgedNotifications, and ui/isWalletConnected
slices with encryption to be added in a separate patch.

Note: `localforage` in itself is a quite good package to interact with
IndexedDB. It has a `localStorage` like API, which means it is simple to use. It
also uses localStorage  as a fallback in case the browser doesn't support
IndexedDB or WebSQL. It also has 1,666,240 downloads and nice documentation.

Issue-URL: https://app.clickup.com/t/865c3zm6w